# backend/docs/swagger.yaml
basePath: /admin-api
definitions:
  github_com_lohasle_nimbus-cloud-framework-go_internal_platform_httpx.Response:
    properties:
      code:
        example: 0
        type: integer
      data: {}
      msg:
        example: ""
        type: string
    type: object
  internal_modules_infra.ConfigSaveRequest:
    properties:
      category:
        type: string
      id:
        type: integer
      key:
        type: string
      name:
        type: string
      remark:
        type: string
      type:
        type: integer
      value:
        type: string
      visible:
        type: boolean
    required:
    - category
    - key
    - name
    type: object
  internal_modules_infra.FileConfigSaveRequest:
    properties:
      config:
        type: string
      id:
        type: integer
      master:
        type: boolean
      name:
        type: string
      remark:
        type: string
      storage:
        type: integer
    required:
    - name
    type: object
...
info:
  contact: {}
  description: Tenant, operations-user and authentication service.
  title: Nimbus Cloud Framework Go System API
  version: "1.0"
paths:
...
  /infra/api-access-log/page:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_lohasle_nimbus-cloud-framework-go_internal_platform_httpx.Response'
      security:
      - BearerAuth: []
      summary: Page API access logs
      tags:
      - Infra Logging
  /infra/config/create:
    post:
      consumes:
      - application/json
      parameters:
      - description: Config
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_modules_infra.ConfigSaveRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_lohasle_nimbus-cloud-framework-go_internal_platform_httpx.Response'
      security:
      - BearerAuth: []
      summary: Create a system parameter
      tags:
      - Infra Config
  /infra/config/delete:
    delete:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_lohasle_nimbus-cloud-framework-go_internal_platform_httpx.Response'
      security:
      - BearerAuth: []
      summary: Delete a system parameter
      tags:
      - Infra Config
  /infra/config/delete-list:
    delete:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_lohasle_nimbus-cloud-framework-go_internal_platform_httpx.Response'
      security:
      - BearerAuth: []
      summary: Delete system parameters in batch
      tags:
      - Infra Config
...
  /infra/config/get:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_lohasle_nimbus-cloud-framework-go_internal_platform_httpx.Response'
      security:
      - BearerAuth: []
      summary: Get a system parameter
      tags:
      - Infra Config
  /infra/config/get-value-by-key:
    get:
      parameters:
      - description: Config key
        in: query
        name: key
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_lohasle_nimbus-cloud-framework-go_internal_platform_httpx.Response'
      summary: Get a public parameter value by key
      tags:
      - Infra Config
  /infra/config/page:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_lohasle_nimbus-cloud-framework-go_internal_platform_httpx.Response'
      security:
      - BearerAuth: []
      summary: Page system parameters
      tags:
      - Infra Config
  /infra/config/update:
    put:
      consumes:
      - application/json
      parameters:
      - description: Config
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_modules_infra.ConfigSaveRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_lohasle_nimbus-cloud-framework-go_internal_platform_httpx.Response'
      security:
      - BearerAuth: []
      summary: Update a system parameter
      tags:
      - Infra Config
  /infra/file-config/create:
    post:
      consumes:
      - application/json
      parameters:
      - description: File config
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_modules_infra.FileConfigSaveRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_lohasle_nimbus-cloud-framework-go_internal_platform_httpx.Response'
      security:
      - BearerAuth: []
      summary: Create a file storage configuration
      tags:
      - Infra File
  /infra/file-config/delete:
    delete:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_lohasle_nimbus-cloud-framework-go_internal_platform_httpx.Response'
      security:
      - BearerAuth: []
      summary: Delete a file storage configuration
      tags:
      - Infra File
  /infra/file-config/delete-list:
    delete:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_lohasle_nimbus-cloud-framework-go_internal_platform_httpx.Response'
      security:
      - BearerAuth: []
      summary: Delete file storage configurations in batch
      tags:
      - Infra File
  /infra/file-config/get:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_lohasle_nimbus-cloud-framework-go_internal_platform_httpx.Response'
      security:
      - BearerAuth: []
      summary: Get a file storage configuration
      tags:
      - Infra File
  /infra/file-config/page:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_lohasle_nimbus-cloud-framework-go_internal_platform_httpx.Response'
      security:
      - BearerAuth: []
      summary: Page file storage configurations
      tags:
      - Infra File
  /infra/file-config/test:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_lohasle_nimbus-cloud-framework-go_internal_platform_httpx.Response'
      security:
      - BearerAuth: []
      summary: Validate a file storage configuration
      tags:
      - Infra File
  /infra/file-config/update:
    put:
      consumes:
      - application/json
      parameters:
      - description: File config
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_modules_infra.FileConfigSaveRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_lohasle_nimbus-cloud-framework-go_internal_platform_httpx.Response'
      security:
      - BearerAuth: []
      summary: Update a file storage configuration
      tags:
      - Infra File
  /infra/file-config/update-master:
    put:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_lohasle_nimbus-cloud-framework-go_internal_platform_httpx.Response'
      security:
      - BearerAuth: []
      summary: Set the primary file storage configuration
      tags:
      - Infra File
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"